[Game] [2048] Fix package naming
