+++ game/quickie2048/src/main/java/com/fakhrirasyids/quickie/game/quickie2048/domain/model/FarthestPosition.kt
@@ -1,0 +1,6 @@
+package com.fakhrirasyids.quickie.game.quickie2048.domain.model
+
+internal data class FarthestPosition(
+    val farthest: Position,
+    val next: Position,
+)
