+++ game/quickie2048/src/main/java/com/fakhrirasyids/quickie/game/quickie2048/domain/model/Position.kt
@@ -1,0 +1,6 @@
+package com.fakhrirasyids.quickie.game.quickie2048.domain.model
+
+data class Position(
+    val x: Int,
+    val y: Int,
+)
