+++ game/quickie2048/src/main/java/com/fakhrirasyids/quickie/game/quickie2048/domain/model/Traversals.kt
@@ -1,0 +1,6 @@
+package com.fakhrirasyids.quickie.game.quickie2048.domain.model
+
+data class Traversals(
+    val x: IntProgression,
+    val y: IntProgression,
+)
