+++ game/quickie2048/src/main/java/com/fakhrirasyids/quickie/game/quickie2048/domain/model/Vector.kt
@@ -1,0 +1,6 @@
+package com.fakhrirasyids.quickie.game.quickie2048.domain.model
+
+data class Vector(
+    val x: Int,
+    val y: Int,
+)
