+++ game/quickie2048/src/main/java/com/fakhrirasyids/quickie/game/quickie2048/ui/model/TileModel.kt
@@ -1,0 +1,11 @@
+package com.fakhrirasyids.quickie.game.quickie2048.ui.model
+
+import androidx.compose.runtime.Immutable
+
+@Immutable
+data class TileModel(
+    val id: String,
+    val curValue: Int,
+    val curPosition: TilePosition,
+    val prevPosition: TilePosition? = null,
+)
