+++ game/quickie2048/src/main/java/com/fakhrirasyids/quickie/game/quickie2048/ui/model/TilePosition.kt
@@ -1,0 +1,9 @@
+package com.fakhrirasyids.quickie.game.quickie2048.ui.model
+
+import androidx.compose.runtime.Immutable
+
+@Immutable
+data class TilePosition(
+    val row: Float,
+    val col: Float,
+)
