+++ game/quickie2048/src/main/java/com/fakhrirasyids/quickie/game/quickie2048/ui/screen/widget/SubHeaderWidget.kt
@@ -1,0 +1,27 @@
+package com.fakhrirasyids.quickie.game.quickie2048.ui.screen.widget
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.material3.Button
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.stringResource
+
+@Composable
+fun SubHeaderWidget(
+    onResetClicked: () -> Unit,
+    modifier: Modifier = Modifier
+) {
+    Row(
+        modifier = modifier.fillMaxWidth(),
+        horizontalArrangement = Arrangement.End,
+    ) {
+        Button(
+            onClick = onResetClicked
+        ) {
+            Text(stringResource(id = com.fakhrirasyids.quickie.baseResource.R.string.sub_header_new_game))
+        }
+    }
+}
