--- game/quickie2048/src/main/java/com/freelanxx/quickie/game/quickie2048/domain/model/FarthestPosition.kt
@@ -1,6 +1,0 @@
-package com.freelanxx.quickie.game.quickie2048.domain.model
-
-internal data class FarthestPosition(
-    val farthest: Position,
-    val next: Position,
-)
