--- game/quickie2048/src/main/java/com/freelanxx/quickie/game/quickie2048/domain/model/Position.kt
@@ -1,6 +1,0 @@
-package com.freelanxx.quickie.game.quickie2048.domain.model
-
-data class Position(
-    val x: Int,
-    val y: Int,
-)
