--- game/quickie2048/src/main/java/com/freelanxx/quickie/game/quickie2048/domain/model/Traversals.kt
@@ -1,6 +1,0 @@
-package com.freelanxx.quickie.game.quickie2048.domain.model
-
-data class Traversals(
-    val x: IntProgression,
-    val y: IntProgression,
-)
