--- game/quickie2048/src/main/java/com/freelanxx/quickie/game/quickie2048/domain/model/Vector.kt
@@ -1,6 +1,0 @@
-package com.freelanxx.quickie.game.quickie2048.domain.model
-
-data class Vector(
-    val x: Int,
-    val y: Int,
-)
