--- game/quickie2048/src/main/java/com/freelanxx/quickie/game/quickie2048/ui/model/TileModel.kt
@@ -1,11 +1,0 @@
-package com.freelanxx.quickie.game.quickie2048.ui.model
-
-import androidx.compose.runtime.Immutable
-
-@Immutable
-data class TileModel(
-    val id: String,
-    val curValue: Int,
-    val curPosition: TilePosition,
-    val prevPosition: TilePosition? = null,
-)
