--- game/quickie2048/src/main/java/com/freelanxx/quickie/game/quickie2048/ui/model/TilePosition.kt
@@ -1,9 +1,0 @@
-package com.freelanxx.quickie.game.quickie2048.ui.model
-
-import androidx.compose.runtime.Immutable
-
-@Immutable
-data class TilePosition(
-    val row: Float,
-    val col: Float,
-)
